+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.1.0 (2023-05-06)
+
+## Features
+
+- Fetch recent casks and formulas from Homebrew repositories
+- Support for GitHub personal access tokens
+
+## Bug Fixes
+
+- Update the repository access logic to handle existing local repositories
+
 
 <!--